Move read to Reconstructor, use one list of Reconstructors in apply_models
